YQL-19747 Restructure commands/interactive module
