+++ ydb/public/lib/ydb_cli/commands/interactive/highlight/color/ya.make
@@ -1,0 +1,11 @@
+LIBRARY()
+
+SRCS(
+    schema.cpp
+)
+
+PEERDIR(
+    contrib/restricted/patched/replxx
+)
+
+END()
