+++ ydb/public/lib/ydb_cli/commands/interactive/highlight/ya.make
@@ -11,9 +11,14 @@ PEERDIR(
     yql/essentials/sql/v1/lexer/antlr4
     yql/essentials/sql/v1/lexer/antlr4_ansi
     yql/essentials/sql/settings
+    ydb/public/lib/ydb_cli/commands/interactive/highlight/color
 )
 
 END()
+
+RECURSE(
+    color
+)
 
 RECURSE_FOR_TESTS(
     ut
