+++ ydb/public/lib/ydb_cli/commands/interactive/highlight/yql_highlighter.h
@@ -2,32 +2,14 @@
 
 #include <contrib/restricted/patched/replxx/include/replxx.hxx>
 
+#include <ydb/public/lib/ydb_cli/commands/interactive/highlight/color/schema.h>
+
 #include <util/generic/fwd.h>
 
 namespace NYdb::NConsoleClient {
 
-    using TColor = replxx::Replxx::Color;
-
     // Colors are provided as for a UTF32 string
     using TColors = replxx::Replxx::colors_t;
-
-    struct TColorSchema {
-        TColor keyword;
-        TColor operation;
-        struct {
-            TColor function;
-            TColor type;
-            TColor variable;
-            TColor quoted;
-        } identifier;
-        TColor string;
-        TColor number;
-        TColor comment;
-        TColor unknown;
-
-        static TColorSchema Monaco();
-        static TColorSchema Debug();
-    };
 
     class IYQLHighlighter {
     public:
