+++ ydb/public/lib/ydb_cli/commands/interactive/ya.make
@@ -14,3 +14,7 @@ PEERDIR(
 
 END()
 
+RECURSE(
+    complete
+    highlight
+)
